% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{property_enum}
\alias{property_enum}
\title{Define a property backed by one or more enum classes}
\usage{
property_enum(..., default = NULL)
}
\arguments{
\item{...}{Enum classes. One or more classes from \code{\link[=new_enum]{new_enum()}}.}

\item{default}{Any. Passed to \code{\link[S7:new_property]{S7::new_property()}}.}
}
\value{
An S7 property typed as the union of \code{...}.
}
\description{
Define a property backed by one or more enum classes
}
