% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{property_union}
\alias{property_union}
\alias{|.S7_property}
\title{Create Property from a Union of Properties}
\usage{
property_union(..., default = NULL)

\method{|}{S7_property}(e1, e2)
}
\arguments{
\item{...}{S7 classes or properties to union.}

\item{default}{Any. Passed to \code{\link[S7:new_property]{S7::new_property()}}.}

\item{e1, e2}{S7 classes or properties to union.}
}
\value{
An S7 property typed as the union of \code{...}.
}
\description{
An analogue to \code{\link[S7:new_union]{S7::new_union()}}. Defines a new property
from multiple
}
\examples{
library(S7)
Piece := new_class(
  properties = list(id = property_union(S7::class_character, S7::class_integer))
)
Piece("a")
Piece(1L)
library(S7)
Piece := new_class(
  properties = list(id = class_string | property_range_discrete(1L, 6L))
)
Piece(3L)
}
