% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{class_string}
\alias{class_string}
\alias{class_integer}
\alias{class_double}
\title{Scalar property presets}
\usage{
class_string

class_integer

class_double
}
\description{
Ready-to-use \code{\link[=property_scalar]{property_scalar()}} variants for the common atomic base
types.
}
