% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{new_enum}
\alias{new_enum}
\title{Create a new enum class}
\usage{
new_enum(name, variants, package = topenv_package_name(parent.frame()))
}
\arguments{
\item{name}{String. Name of the new class.}

\item{variants}{Character vector. Allowed values.}

\item{package}{String or NULL. Package to attribute the class to.
Defaults to the caller's package, if any.}
}
\value{
An S7 class generator that inherits from \link{Enum}.
}
\description{
Create a new enum class
}
\examples{
GridShape <- new_enum("GridShape", c("Square", "Hexagon"))
GridShape("Square")
}
