% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{property_scalar}
\alias{property_scalar}
\title{Define a scalar property}
\usage{
property_scalar(class, default = NULL)
}
\arguments{
\item{class}{S7 class spec. Base type the value must be an instance of.}

\item{default}{Any. Passed to \code{\link[S7:new_property]{S7::new_property()}}.}
}
\value{
An S7 property whose value must be a scalar atomic.
}
\description{
Define a scalar property
}
