% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7x-package.R
\docType{package}
\name{s7x-package}
\alias{s7x}
\alias{s7x-package}
\title{s7x: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
